requirements.txt
# torch==2.7.0 torchvision==0.22.0 torchaudio==2.7.0
# Download the appropriate PyTorch for your device
transformers==4.57.1
huggingface_hub==0.36.2
hf_xet==1.2.0
qwen_vl_utils==0.0.14
pyyaml==6.0.3
numpy==2.2.6
pillow==12.1.0
tqdm==4.67.3
scipy==1.15.3
colorama==0.4.6
peft==0.18.1
einops==0.8.2
timm==1.0.24
triton==2.1.0
decord==0.6.0
hydra-core==1.3.2
iopath==0.1.10
nncore==0.4.7
pycocotools==2.0.10
pydantic==2.11.7
pysrt==1.1.2
tensordict==0.9.1
pandas==2.3.1
termplotlib==0.3.9
